Support postman Oauth 1.0 and Oauth 2.0 config
This commit adds support for importing as much configuration as we can
from Postman v2.0 and v2.1 exports with Oauth 1.0 and 2.0
authentication.

Note: For Oauth2.0 this means setting authorization to Oauth2 but
generating a disabled config, since postman does not export anything we
can use to fill our config values.

Fixes #1240