Optionally make the prompt plugin pre-fill the last entered value
Added an option to the prompt plugin to save the last entered value and have that set as default value the next time the prompt shows. This saves you from having to enter all details for each request while still being able to change them if needed. This should be BC and therefore the option is ignored when a storage key is set explicitly; this will still return the cached value directly.