[Bug Fix 1288] "Undefined" for nested variables (#1637)
* Update nunjucks-tags.js

(Fixes bug #1288) Extract parent from nunjucks tag when it is a nested property nested. The parent is then used to fetch the keyContext

* Revert "Update nunjucks-tags.js"

This reverts commit 4d4da99de3220ee38632081bc405d964a468fc4c.

* Update render.js

Added function that recursively gets the keys. So they will show the correct parent location in the nunjuck tooltip.

* Rework after review from gschier

A few little changes after review of the initial change.

* Added keySource for Arrays

Added loop to store Key and Source when item is an Array.