Add "When Expired" setting for request resend trigger condition.
This adds a new trigger condition for re-sending a request for the
`response` plugin called "When Expired". This condition allows
specifying a maximum age of a response to use. This is very useful when
configuring a hierarchy of environment variables that ultimately depend
on an auth token that needs to be refreshed periodically.

Previous workarounds are:
1. Use "always refresh". This fails in cases where you must use the same
   auth token since it is refreshed for every request.
2. Use "no history" or "never" and manually refresh the token. This is
   inconvenient and requires pinning the token.

Alternative solutions are:
1. Add more sophisticated logic to determine when the token needs to be
   refreshed (e.g. refresh if the current request returns a `401` status
   code)
2. Write a custom response plugin to time out the token. I believe this
   functionality is small enough and general enough to be contributed
   back generally.

Fixes #1972