Fix curl importers when input contains escaped array params
In Chrome on Windows, the `Copy as cURL (bash)` action is escaping array query params.
We are getting the following url: `'http://example.com/?a\[0\]=1&a\[1\]=1'`

Unfortunately, the importers will parse the params without unescaping the keys, which will create incorrect
params:

```
{
  "disabled": false,
  "name": "a\[0\]",
  "value": "1"
}
```

To fix it, I did a simple change replacing any `\[ \]` into non-escaped value `[ ]` in the key name.

Closes #1963