Change it block from lambda to anonymous function
Changing the function passed to the it block from a lambda to an anonymous function allows access to the mocha context from within the test. This allows users to change the timeouts for individual unit tests via the exposed code blocks.
ex:

this.timeout(10000);
const response1 = await insomnia.send();
expect(response1.status).to.equal(200);

will timeout after 10 seconds instead of 5.