adds `-- TSCONVERSION` per phase 2 so we know what to target for phase 3
This was done by one-by-one looking at the results of the search `ts-expect-error(?= [^-])` and the replace `ts-expect-error -- TSCONVERSION` in `insomnia-app`.

The idea with this convention, which we followed everywhere else, is that for Phase 3 we will either fix (and therefore remove) or remove the `-- TSCONVERSION ` part of each of these ignores.