removes references to old build artifact location
well, except for the .gitignore.  I guess I could remove it from the .gitignore, commit it, commit deleting it, which would clear it from everyone's projects... but I'll just leave it there for a bit in the .gitignore.  hopefully people either start fresh or run `clean` from the root every once in a while, which will clean this up.

After a few weeks or months we can remove these two references (in the package.json.clean and .gitignore) altogether if we feel like it.