Fixes clean scripts (#3400)
* removes (unused and unwanted) husky scripts

might as well, it's only taking up space for nothing in our node_modules

* adds `lerna run clean` to main clean script

we didn't have `clean` scripts in the packages before, but now that we do, it makes sense to.... hahah... _use them_.

Note that it's important that we run the package scripts first, because otherwise after the rimraf (which is now at the end) there's no `lerna` left to call.

* adds missing clean script to insomnia-components

although it looks like there are other changes at first glance, there shouldn't be any behavioral differences now except the fact that build will always generate SVGs (which, was an oversight/bug that it wasn't before).  We need this to occur so that when we clean it removes the things that were built, and when we build it replaces those same things.

* adds a postclean script to insomnia-app like every other package

I'll be honest, I'm not entirely sure what the reason is for having `postclean` scripts in every package with (mostly) `rimraf dist` (i.e. instead of just including that in the main clean script at the end), but that's clearly the convention - so for now I'm just following it

* removes references to old build artifact location

well, except for the .gitignore.  I guess I could remove it from the .gitignore, commit it, commit deleting it, which would clear it from everyone's projects... but I'll just leave it there for a bit in the .gitignore.  hopefully people either start fresh or run `clean` from the root every once in a while, which will clean this up.

After a few weeks or months we can remove these two references (in the package.json.clean and .gitignore) altogether if we feel like it.

* removes unused reference in config to build artifact

I searched high and low for every place I could find that consumes this file and none of them seem to use this field anyway.  And even if they did, it's wrong.  In fact, it's been wrong for a while I'd guess because (relative to the root, if that's what it was intended to mean).

As far as I can tell this was added https://github.com/Kong/insomnia/commit/c6a7c4d68272b3c4f3d78d4dc9031ae8e49b3f7a#diff-6918d2464ad7fb0f3e985d8587ce0c4218eeba6c6dea60c58642251f4db99ecb

So I even checked out that commit and looked at all usages of that file at that time and it was, there too, unused (so far as I can tell).  Appears to be a case of wishful thinking that it'd be useful for something, not sure.