clears errors in tests
these specific tests were failing after changes in this PR and I'm trying to figure out why - clearing the errors first helps a lot with that.

I rearranged the commits to make this one first so that I could verify that these changes do not affect the prior results of the tests (and, they do not).