removes storybook (and related dependencies) from insomnia-app
We noticed yesterday that these files are no longer used.  If, in the future, when spending more attention on insomnia-components, we want them, then they'll be safe and sound here in git.  Until then there's no need to keep them around and depend on them.

Some of them, anyway, were already moved over e.g. `help-tooltip`, and `tooltip` and are duplicates as they exist in insomnia-app.