Make Swagger 2 importer generate deterministic IDs
This is a backport of a feature that was added to the OpenAPI 3.0
importer two years ago in 3e2311e4. As it is now, the placeholder ID
gets replaced by a UUIDv4 identifier, which is pseudorandom and
therefore not deterministic. This is problematic because the importer
will not have any obvious way to determine identity between requests,
and so it duplicates them every time.

Fixes INS-567.