changes default project id value
we already changed this once this week.  my understanding is that from running instance of the app to another running instance of the app it doesn't matter if this is ever changed, but I wanted to separate it out in its own commit just to call it out to be doubly sure.