use `omit in toc` directive
I checked out the docs for the TOC generator (https://marketplace.visualstudio.com/items?itemName=yzhang.markdown-all-in-one#table-of-contents) and saw that we can disable some headers individually.

It seemed like Configuration.Options and Configuration.Scripts were good ones to have around so I left those, but lemme know if not.