removes duplicate logic and fixes render skipping
I noticed a bug that didn't show up before because there was previously no way to beautify (and, thus, trigger this code path) for documents that weren't showing the beautify button.  Now, since the keyboard shortcut will trigger it for everything, what can happen is that the viewport and cursor will reset even if there's nothing to change.  The `_handleBeautify` method before doesn't serve much purpose since all it does is call `_prettify` (and, it was the only thing calling `_prettify`).  Before doing that, though, it's worthless to bother setting the value.  This way, if the document cannot be prettified then it won't be.