rebases functionality
in response-viewer: I had to extract _getMode in order to use it in the click link handler.  To do that, though, I needed to also extract _getBody.  As I did that, I saw that character-for-character code for creating the body was duplicated all around the file.  So I removed that duplication while I was here (along with reduced the else-return complexity).  This then illuminated that the only place _decodeIconv was being called was (now) in the _getBody method, which meant that it could be inlined.