fixes type error in tests.
I was on my way to writing some tests to cover the things addressed in this PR.  I was happy to see that we already covered what I was planning to test.  However, there were some very surface-level type errors in the tests so I thought it appropriate to fix them while I was in this area of the code for this job.