fixes error handling (and removes some else-returns) in pkg.ts
prior to this, the `reject` on line 28 does not hault execution.  it's reasonable that this mistake was made because it's an easy one to make if you don't allow the function that ultimately had the failure case to have the control flow.

As it was before, code _after_ the reject was still executing.

Now, the promise rejection is handled in the place it really occurs.  Promises are designed to work well with `throw` like this and since there is no other callback that the throw is captured in (e.g. like a setTimeout or something) this is totally fine to do.  Here is what the output would look like:

```
> ts-node src/scripts/pkg.ts

Error: [pkg-inso] Unsupported OS: freebsd
    at getTargets (/home/dimitri/src/github.com/Kong/insomnia/packages/insomnia-inso/src/scripts/pkg.ts:24:9)
    at /home/dimitri/src/github.com/Kong/insomnia/packages/insomnia-inso/src/scripts/pkg.ts:29:21
    at new Promise (<anonymous>)
    at pkg (/home/dimitri/src/github.com/Kong/insomnia/packages/insomnia-inso/src/scripts/pkg.ts:28:10)
    at Object.<anonymous> (/home/dimitri/src/github.com/Kong/insomnia/packages/insomnia-inso/src/scripts/pkg.ts:67:1)
    at Module._compile (internal/modules/cjs/loader.js:1137:30)
    at Module.m._compile (/home/dimitri/src/github.com/Kong/insomnia/packages/insomnia-inso/node_modules/ts-node/src/index.ts:1056:23)
    at Module._extensions..js (internal/modules/cjs/loader.js:1157:10)
    at Object.require.extensions.<computed> [as .ts] (/home/dimitri/src/github.com/Kong/insomnia/packages/insomnia-inso/node_modules/ts-node/src/index.ts:1059:12)
    at Module.load (internal/modules/cjs/loader.js:985:32)
```