moves insomnia config filename to resolver and splits up error
otherwise, like it was before, you virtually always see a console.error for a case that isn't really an error (a file that's not expected to always be present was not found) when the intent for why we added that console.error in the first place was for parsing errors.

also, now that we have the error narrowed down quite a bit more, it's easier to give it a more informative message.