add filename in one place, only
before this, the config file was being added in multiple places, and that created a scenario where (on line 33 of settings.ts, for example) one was missed.  To fix this, we can just add the file path at the beginning, and then never need to fuss with it again in subsequent usages and logging and error reporting.