updates type-fest to latest
we were previously blocked on this upgrade by our node version: but now that we're up to date we can be on latest.  there are a few things that have happened in the last year in this library that I want to use from time to time but couldn't.  should be no breaking changes (they majored because they dropped lower node and ts versions).