4688 Header ending with -bin need buffer values
For metadata entries where the name ends with -bin the values must be
Buffers instead of strings.

The solucion checks if the metadata entry should be a buffer and, when
yes, constructs a Buffer from the value. The value is split into the
encoding part and the real value part. If no encoding is specified the
Buffer will be constructed with the default encoding.