feat: WIP generate insomnia-inso docs from source
Add a script that generates the reference docs for insomnia-inso from
the source code. Commander exposes commands, options and descriptions
which we can use to generate the markdown files.

This is a WIP to start a convo and an initial implementation, the script
doesn't work as it is because `cli.ts` doesn't export `program` and
`version`.