[INS-1234] Include token as backup credential when initially fetching repo contents
The 'custom' Git tab was failing because it would send the credentials using the token field while the submit method expected it in the password field. Since the method is reused by multiple code paths the most resilient patch is simply to add the alternative possibility, the token, as a fallback.